removed test/utils.py and placed the function within it into janitor/utils.py
